Added more failing tests.
