+++ previous_years/test/test_utils.py
@@ -15,6 +15,10 @@ from core.test.test_base import RequestFactoryBase
 
 from costcentre.test.factories import ArchivedCostCentreFactory
 
+from forecast.models import (
+    FinancialPeriod,
+)
+
 from previous_years.models import (
     ArchivedFinancialCode,
     ArchivedForecastData,
@@ -142,3 +146,10 @@ class DownloadPastYearForecastSetup(TestCase, RequestFactoryBase):
         can_view_forecasts = Permission.objects.get(codename="can_view_forecasts")
         self.test_user.user_permissions.add(can_view_forecasts)
         self.test_user.save()
+
+
+def hide_adjustment_columns(self):
+    for period_code in range(13,16):
+        financial_period_obj = FinancialPeriod.objects.get(financial_period_code=period_code)
+        financial_period_obj.display_figure = False
+        financial_period_obj.save()
