+++ previous_years/test/test_view_previous_year_project.py
@@ -11,7 +11,11 @@ from forecast.views.view_forecast.project_details import (
     GroupProjectDetailsView,
 )
 
-from previous_years.test.test_utils import DownloadPastYearForecastSetup
+from previous_years.test.test_utils import (
+    DownloadPastYearForecastSetup,
+    hide_adjustment_columns
+)
+
 
 TOTAL_COLUMN = -5
 SPEND_TO_DATE_COLUMN = -2
@@ -129,3 +133,9 @@ class ViewProjectDetailsTest(DownloadPastYearForecastSetup):
             period=self.archived_year,
         )
         self.check_response(resp)
+
+
+class ViewProjectDetailsAdjustmentColumnsTest(ViewProjectDetailsTest):
+    def setUp(self):
+        super().setUp()
+        hide_adjustment_columns()
